+++ src/Symfony/Component/Serializer/CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 ---
 
  * Add `XmlEncoder::SAVE_OPTIONS` context option
+ * Deprecate `MissingConstructorArgumentsException` in favor of `MissingConstructorArgumentException`
 
 6.2
 ---